Wrap all git mirrors functionality in experiment

+++ bootstrap/integration/bootstrap_tester.go
@@ -17,6 +17,8 @@ import (
 	"syscall"
 	"testing"
 
+	"github.com/buildkite/agent/experiments"
+
 	"github.com/buildkite/bintest"
 )
 
@@ -66,11 +68,6 @@ func NewBootstrapTester() (*BootstrapTester, error) {
 		return nil, err
 	}
 
-	gitMirrorsDir, err := ioutil.TempDir("", "bootstrap-git-mirrors")
-	if err != nil {
-		return nil, err
-	}
-
 	repo, err := createTestGitRespository()
 	if err != nil {
 		return nil, err
@@ -84,7 +81,6 @@ func NewBootstrapTester() (*BootstrapTester, error) {
 			"HOME=" + homeDir,
 			"BUILDKITE_BIN_PATH=" + pathDir,
 			"BUILDKITE_BUILD_PATH=" + buildDir,
-			"BUILDKITE_GIT_MIRRORS_PATH=" + gitMirrorsDir,
 			"BUILDKITE_HOOKS_PATH=" + hooksDir,
 			"BUILDKITE_PLUGINS_PATH=" + pluginsDir,
 			`BUILDKITE_REPO=` + repo.Path,
@@ -102,11 +98,25 @@ func NewBootstrapTester() (*BootstrapTester, error) {
 			`BUILDKITE_JOB_ID=1111-1111-1111-1111`,
 			`BUILDKITE_AGENT_ACCESS_TOKEN=test`,
 		},
-		PathDir:       pathDir,
-		BuildDir:      buildDir,
-		HooksDir:      hooksDir,
-		PluginsDir:    pluginsDir,
-		GitMirrorsDir: gitMirrorsDir,
+		PathDir:    pathDir,
+		BuildDir:   buildDir,
+		HooksDir:   hooksDir,
+		PluginsDir: pluginsDir,
+	}
+
+	// Support testing experiments
+	if exp := experiments.Enabled(); len(exp) > 0 {
+		bt.Env = append(bt.Env, `BUILDKITE_AGENT_EXPERIMENT=`+strings.Join(exp, ","))
+
+		if experiments.IsEnabled(`git-mirrors`) {
+			gitMirrorsDir, err := ioutil.TempDir("", "bootstrap-git-mirrors")
+			if err != nil {
+				return nil, err
+			}
+
+			bt.GitMirrorsDir = gitMirrorsDir
+			bt.Env = append(bt.Env, "BUILDKITE_GIT_MIRRORS_PATH="+gitMirrorsDir)
+		}
 	}
 
 	// Windows requires certain env variables to be present
@@ -329,8 +339,10 @@ func (b *BootstrapTester) Close() error {
 	if err := os.RemoveAll(b.PluginsDir); err != nil {
 		return err
 	}
-	if err := os.RemoveAll(b.GitMirrorsDir); err != nil {
-		return err
+	if b.GitMirrorsDir != "" {
+		if err := os.RemoveAll(b.GitMirrorsDir); err != nil {
+			return err
+		}
 	}
 	return nil
 }
